+++ obscurepy/handlers/__init__.py
@@ -2,4 +2,5 @@ from obscurepy.handlers.assign_handler import AssignHandler
 from obscurepy.handlers.call_handler import CallHandler
 from obscurepy.handlers.classdef_handler import ClassDefHandler
 from obscurepy.handlers.functiondef_handler import FunctionDefHandler
-from obscurepy.handlers.literal_handler import LiteralHandler
+from obscurepy.handlers.name_handler import NameHandler
+from obscurepy.handlers.constant_handler import ConstantHandler