+++ tests/test_complete_chain.py
@@ -11,7 +11,7 @@ class CompleteChainTest(unittest.TestCase):
         self.tracker = DefinitionTracker.get_instance()
         self.source = 'class FirstClass:\n\tpass\n\n' \
                       'class SecondClass:\n\tpass\n\n' \
-                      'def FirstFunction():\n\tc = 42\n\n' \
+                      'def FirstFunction():\n\tc = "a literal"\n\n' \
                       'def SecondFunction():\n\tpass\n\n' \
                       'a = FirstClass()\n\n' \
                       'b = SecondClass()\n\n' \